Refactor graphics test framework
They are now individually debuggable using external tools
Using the GFX_TESTS_DEBUG=1 environment variable, they run inside a window loop
